+++ geometry/segment.hh
@@ -11,10 +11,10 @@ class Segment {
     public :
 	Segment(Vertex u, Vertex v);
 	float length();
-	int width();
-	int height();
+	float width();
+	float height();
 	Segment reduce(float value);
-	Vertex randomPos(int seed, int n, int a, int b); // Renvoir un vertex sur le segment [u,v], à une position entre a% and b%.
+	Vertex randomPos(int seed, int n, float a, float b); // Renvoie un vertex sur le segment [u,v], à une position entre a et b.
 };
 
 #endif
