+++ main.cpp
@@ -1,9 +1,12 @@
 #include "all_includes.hh"
 
+// TODO : créer les routes dans les bâtiments
+// TODO : faire des enum SetCoin, SetCardinal, SetSommetTriangle, SetCoteTriangle.
+
 int main() {
 	// Générer une tile de base
 	std::cout << "Initial seed = " << Chose::initialSeed << std::endl;
-	int size = 20000;
+	float size = 20000;
 	Vertex ne(size, size, 0);
 	Vertex se(size, 0, 0);
 	Vertex sw(0, 0, 0);
